--- a/aggregators/internal/hdrhistogram/hdrhistogram.go
+++ b/aggregators/internal/hdrhistogram/hdrhistogram.go
@@ -25,6 +25,7 @@
 // THE SOFTWARE.
 
 // Package hdrhistogram provides an optimized histogram for sparse samples.
+// This is a stop gap measure until we have [packed histogram implementation](https://www.javadoc.io/static/org.hdrhistogram/HdrHistogram/2.1.12/org/HdrHistogram/PackedHistogram.html).
 package hdrhistogram
 
 import (