--- a/aggregators/internal/hdrhistogram/hdrhistogram_test.go
+++ b/aggregators/internal/hdrhistogram/hdrhistogram_test.go
@@ -5,6 +5,7 @@
 package hdrhistogram
 
 import (
+	"math"
 	"math/rand"
 	"testing"
 
@@ -31,6 +32,41 @@ func TestMerge(t *testing.T) {
 	expectedSnap := hist1.Export()
 
 	assert.Empty(t, cmp.Diff(expectedSnap, histRep1.getHDRSnapshot()))
+}
+
+func TestBuckets(t *testing.T) {
+	buckets := func(h *hdrhistogram.Histogram) (int64, []int64, []float64) {
+		distribution := h.Distribution()
+		counts := make([]int64, 0, len(distribution))
+		values := make([]float64, 0, len(distribution))
+
+		var totalCount int64
+		for _, b := range distribution {
+			if b.Count <= 0 {
+				continue
+			}
+			count := int64(math.Round(float64(b.Count) / histogramCountScale))
+			counts = append(counts, count)
+			values = append(values, float64(b.To))
+			totalCount += count
+		}
+		return totalCount, counts, values
+	}
+	hist := getTestHistogram()
+	histRep := New()
+
+	for i := 0; i < 1_000_000; i++ {
+		v := rand.Int63n(3_600_000_000)
+		c := rand.Int63n(1_000)
+		hist.RecordValues(v, c)
+		histRep.RecordValues(v, c)
+	}
+	actualTotalCount, actualCounts, actualValues := histRep.Buckets()
+	expectedTotalCount, expectedCounts, expectedValues := buckets(hist)
+
+	assert.Equal(t, expectedTotalCount, actualTotalCount)
+	assert.Equal(t, expectedCounts, actualCounts)
+	assert.Equal(t, expectedValues, actualValues)
 }
 
 func getTestHistogram() *hdrhistogram.Histogram {