The PR improves bucket calculation for HDR histogram representation by utilizing the sparse representation we use. The buckets are calculated for only the buckets that are used for recording values and the allocations are also made accordingly.