@lahsivjar 2023-07-07
Benchmarking results with main:

```
name                        old time/op    new time/op    delta
CombinedMetricsEncoding-10    5.68µs ± 2%    5.65µs ± 2%     ~     (p=0.363 n=10+10)
CombinedMetricsDecoding-10    6.31µs ± 3%    6.55µs ± 2%   +3.82%  (p=0.000 n=10+10)
CombinedMetricsToBatch-10      739µs ± 0%     350µs ± 1%  -52.61%  (p=0.000 n=7+10)
EventToCombinedMetrics-10     1.24µs ± 3%    1.19µs ± 3%   -3.73%  (p=0.000 n=9+10)

name                        old alloc/op   new alloc/op   delta
CombinedMetricsEncoding-10    2.77kB ± 0%    2.77kB ± 0%     ~     (all equal)
CombinedMetricsDecoding-10    13.1kB ± 0%    13.1kB ± 0%     ~     (p=1.260 n=9+10)
CombinedMetricsToBatch-10     4.25MB ± 0%    0.04MB ± 0%  -99.06%  (p=0.000 n=10+10)
EventToCombinedMetrics-10     3.42kB ± 0%    3.42kB ± 0%     ~     (all equal)

name                        old allocs/op  new allocs/op  delta
CombinedMetricsEncoding-10      79.0 ± 0%      79.0 ± 0%     ~     (all equal)
CombinedMetricsDecoding-10      80.0 ± 0%      80.0 ± 0%     ~     (all equal)
CombinedMetricsToBatch-10        661 ± 0%       339 ± 0%  -48.71%  (p=0.000 n=10+10)
EventToCombinedMetrics-10       19.0 ± 0%      19.0 ± 0%     ~     (all equal)
```