@lahsivjar 2023-07-07
Mem profile for `CombinedMetricsToBatch`:

### Before: 
<img width="699" alt="Screenshot 2023-07-07 at 5 56 01 PM" src="https://github.com/elastic/apm-aggregation/assets/1427601/f9be4cb3-24c3-44a6-a316-3193fb6eaa0b">

### After:
<img width="911" alt="Screenshot 2023-07-07 at 5 56 14 PM" src="https://github.com/elastic/apm-aggregation/assets/1427601/5590aab1-5288-404a-8bcf-3b5c1dadf6fa">